refactor(map,plotly): extract csvExport, snapping module, and useSnapping hook
